refactored naming and page location
